Sort_Images/pyproject.toml
[project]
name = "sort-images"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pillow>=12.1.1",
]
